requirements.txt
# flow
Pillow==7.0.0
numpy==1.21.1
tqdm==4.42.1
timm==0.4.12
albumentations==1.0.3
imgaug==0.4.0
adabelief-pytorch==0.2.1
opencv-python==4.5.5.62
pandas==1.3.5

# etl
prefect==0.14.0

# logging
sentry-sdk==0.19.5
pymongo==4.0.1

# worker
celery==5.0.5

# s3
minio==7.1.0